requirements.txt
# sub-converter-flask/requirements.txt

Flask
requests
PyYAML
gunicorn